[JAP] chiển dữ liệu từ main sang chi tiết sản phẩm

+++ WPK_2/src/main/java/com/poly/repository/SanphamDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poly.entities.SanPham;
 
-public interface SanphamDAO extends JpaRepository<SanPham, Long> {
+public interface SanphamDAO extends JpaRepository<SanPham, Integer> {
 
 }
